Remove duplicate code in tests
