+++ test/integration/ggrc/__init__.py
@@ -11,11 +11,14 @@ import csv
 from StringIO import StringIO
 
 from sqlalchemy import exc
+from sqlalchemy import func
+from sqlalchemy.sql.expression import tuple_
 from flask.ext.testing import TestCase as BaseTestCase
 
 from ggrc import db
 from ggrc.app import app
 from integration.ggrc.api_helper import Api
+from integration.ggrc.models import factories
 
 # Hide errors during testing. Errors are still displayed after all tests are
 # done. This is for the bad request error messages while testing the api calls.
@@ -297,3 +300,37 @@ class TestCase(BaseTestCase, object):
     """Assert slugs for each date format ent datetime"""
     for date_string in self.generate_date_strings(datetime_value, formats):
       self.assert_slugs(alias, date_string, slugs)
+
+  @staticmethod
+  def _get_latest_object_revisions(objects):
+    """Get latest revisions of given objects."""
+    object_tuples = [(o.id, o.type) for o in objects]
+    from ggrc.models import Revision
+    revisions = Revision.query.filter(
+        Revision.id.in_(
+            db.session.query(func.max(Revision.id)).filter(
+                tuple_(
+                    Revision.resource_id,
+                    Revision.resource_type,
+                ).in_(object_tuples)
+            ).group_by(
+                Revision.resource_type,
+                Revision.resource_id,
+            )
+        )
+    )
+    return revisions
+
+  def _create_snapshots(self, audit, objects):
+    """Create snapshots of latest object revisions for given objects."""
+    revisions = self._get_latest_object_revisions(objects)
+    snapshots = [
+        factories.SnapshotFactory(
+            child_id=revision.resource_id,
+            child_type=revision.resource_type,
+            revision=revision,
+            parent=audit,
+        )
+        for revision in revisions
+    ]
+    return snapshots
